More chances to confrom to new protocol

+++ src/config.rs
@@ -74,7 +74,7 @@ pub struct CameraConfig {
 
     #[validate(range(min = 0, max = 31, message = "Invalid channel", code = "channel_id"))]
     #[serde(default = "default_channel_id")]
-    pub channel_id: u32,
+    pub channel_id: u8,
 }
 
 #[derive(Debug, Deserialize, Validate, Clone)]
